+++ __tests__/component-remove.js
@@ -1,21 +1,19 @@
 import { Component, DI } from '../src/shamUI';
 import { renderComponent, renderApp } from './helpers';
 
-it( 'remove', async() => {
-    expect.assertions( 1 );
-    await renderComponent( Component );
-    DI.resolve( 'sham-ui' ).render.unregister( 'dummy' );
-    const store = DI.resolve( 'sham-ui:store' );
+it( 'remove', () => {
+    renderComponent( Component );
+    const { store } = DI.resolve( 'sham-ui' );
+    store.findById( 'dummy' ).remove();
     expect( store.findById( 'dummy' ) ).toBe( undefined );
 } );
 
-it( 'remove one', async() => {
-    expect.assertions( 2 );
-    DI.bind( 'component-binder', () => {
+it( 'remove one', () => {
+    renderApp( () => {
         class SelfDestroyedComponent extends Component {
             render() {
                 super.render( ...arguments );
-                this.UI.render.unregister( this.ID );
+                this.remove();
             }
         }
         new SelfDestroyedComponent( {
@@ -27,16 +25,7 @@ it( 'remove one', async() => {
             container: document.querySelector( 'body' )
         } );
     } );
-    await renderApp();
     const store = DI.resolve( 'sham-ui:store' );
     expect( store.findById( 'first' ) ).toBeUndefined();
     expect( store.findById( 'second' ).ID ).toBe( 'second' );
-} );
-
-it( 'remove non exists component', async() => {
-    expect.assertions( 1 );
-    await renderComponent( Component );
-    DI.resolve( 'sham-ui' ).render.unregister( 'non-exists' );
-    const store = DI.resolve( 'sham-ui:store' );
-    expect( store.findById( 'dummy' ) ).toBeInstanceOf( Component );
 } );