+++ __tests__/resolve-container.js
@@ -1,36 +1,32 @@
-import { DI, Component } from '../src/shamUI';
+import { Component } from '../src/shamUI';
 import { renderComponent, renderApp } from './helpers';
 
-it( 'bindEvents', async() => {
-    expect.assertions( 1 );
+it( 'bindEvents', () => {
     class Dummy extends Component {
         bindEvents() {
             expect( this.container ).toBeInstanceOf( Element );
         }
     }
-    await renderComponent( Dummy );
+    renderComponent( Dummy );
 } );
 
-it( 'render', async() => {
-    expect.assertions( 1 );
+it( 'render', () => {
     class Dummy extends Component {
         render() {
             expect( this.container ).toBeInstanceOf( Element );
         }
     }
-    await renderComponent( Dummy );
+    renderComponent( Dummy );
 } );
 
-it( 'empty component', async() => {
-    expect.assertions( 1 );
+it( 'empty component', () => {
     class Dummy extends Component {}
     let component;
-    DI.bind( 'component-binder', () => {
+    renderApp( () => {
         component = new Dummy( {
             ID: 'dummy',
             container: document.querySelector( 'body' )
         } );
     } );
-    await renderApp();
     expect( component.container ).toBeInstanceOf( Element );
 } );