+++ __tests__/store.js
@@ -1,56 +1,49 @@
 import { DI, Component } from '../src/shamUI';
 import { renderApp } from './helpers';
 
-it( 'find', async() => {
-    expect.assertions( 2 );
+it( 'find', () => {
     let foo;
-    DI.bind( 'component-binder', () => {
+    renderApp( () => {
         foo = new Component( {
             ID: 'foo',
             container: document.querySelector( 'body' )
         } );
     } );
-    await renderApp();
     const store = DI.resolve( 'sham-ui:store' );
     expect( store.find( component => component.ID === 'foo' ) ).toEqual( foo );
     expect( store.find( component => component.ID === 'bar' ) ).toBe( null );
 } );
 
-it( 'filter', async() => {
-    expect.assertions( 2 );
+it( 'filter', () => {
     let foo;
-    DI.bind( 'component-binder', () => {
+    renderApp( () => {
         foo = new Component( {
             ID: 'foo',
             container: document.querySelector( 'body' )
         } );
     } );
-    await renderApp();
     const store = DI.resolve( 'sham-ui:store' );
     const components = store.filter( component => component.ID === 'foo' );
     expect( components ).toHaveLength( 1 );
     expect( components[ 0 ] ).toEqual( foo );
 } );
 
-it( 'map', async() => {
-    expect.assertions( 2 );
-    DI.bind( 'component-binder', () => {
+it( 'map', () => {
+    renderApp( () => {
         new Component( {
             ID: 'foo',
             container: document.querySelector( 'body' )
         } );
     } );
-    await renderApp();
     const store = DI.resolve( 'sham-ui:store' );
     const ids = store.map( component => component.ID );
     expect( ids ).toHaveLength( 1 );
     expect( ids[ 0 ] ).toEqual( 'foo' );
 } );
 
-it( 'forEachId', async() => {
-    expect.assertions( 6 );
+it( 'forEachId', () => {
     let foo, bar;
-    DI.bind( 'component-binder', () => {
+    renderApp( () => {
         foo = new Component( {
             ID: 'foo',
             container: document.querySelector( 'body' )
@@ -60,7 +53,6 @@ it( 'forEachId', async() => {
             container: document.querySelector( 'body' )
         } );
     } );
-    await renderApp();
     const allComponents = jest.fn();
     DI.resolve( 'sham-ui:store' ).forEachId( [ 'foo', 'bar' ], allComponents );
     expect( allComponents ).toHaveBeenCalledTimes( 2 );