+++ src/shamUI.js
@@ -1,13 +1,11 @@
 import './initializer';
 
-export { FSM as Fsm, states as FsmStates } from './engine';
-export { default as ComponentStore } from './engine/store';
-export { default as DI, inject } from './DI';
+import Render from './engine';
+import Store from './engine/store';
+export { default as DI } from './DI';
 export { default as Component } from './component';
 export { default as options } from './options/decorator';
 export { default as configureOptions } from './options/helper';
-export { Fsm as FsmDefault } from './fsm';
-export { default as FsmState } from './fsm/state';
 
 import DI from './DI';
 
@@ -18,15 +16,10 @@ export default class ShamUI {
 
     /**
      * Создать экземпляр
-     * @param {Boolean} autoStart
      */
-    constructor( autoStart = false ) {
+    constructor() {
         DI.bind( 'sham-ui', this );
-        const Fsm = DI.resolve( 'fsm' );
-        this.render = new Fsm();
-        this.render.run();
-        if ( autoStart ) {
-            this.render.ALL();
-        }
+        this.store = new Store();
+        this.render = new Render();
     }
 }